refactor: new IAM token authentication method  to create buckets.
There is no need to auto-create SA for the module work now. But your still can specify Existing SA.
Be careful with bucket policy with the IAM token authentication method for Terraform (without existing service account). It is not possible to add a rule for the IAM token to the bucket policy. You can add a statement for IAM user or group who run Terraform.